requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
pyrealsense2>=2.50.0
pyyaml>=5.4.0
scipy>=1.7.0
